Item Rendering Hack
Now all but the resource crop seeds render in the inventory.
