+++ src/main/java/com/InfinityRaider/AgriCraft/api/v1/IAgriCraftSeed.java
@@ -6,7 +6,7 @@ import net.minecraftforge.common.IPlantable;
 import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 
-public interface IAgriCraftSeed extends IPlantable, IconRegisterable {
+public interface IAgriCraftSeed extends IAgriCraftRenderable, IPlantable {
     /** Gets the plant for this seed */
     public IAgriCraftPlant getPlant();
 
@@ -17,7 +17,4 @@ public interface IAgriCraftSeed extends IPlantable, IconRegisterable {
     @SideOnly(Side.CLIENT)
     public String getInformation();
 
-    /** Gets the registered TextureAtlasSprite icon for this seed */
-    @SideOnly(Side.CLIENT)
-    TextureAtlasSprite getIcon(ItemStack stack);
 }
